refactor: getFootprintInfo함수 WoofRepository로 옮기기
