+++ android/app/src/main/java/com/happy/friendogly/application/di/AppModule.kt
@@ -7,7 +7,6 @@ import com.happy.friendogly.crashlytics.CrashlyticsHelper
 import com.happy.friendogly.data.repository.AddressRepositoryImpl
 import com.happy.friendogly.data.repository.AuthRepositoryImpl
 import com.happy.friendogly.data.repository.ClubRepositoryImpl
-import com.happy.friendogly.data.repository.FootprintRepositoryImpl
 import com.happy.friendogly.data.repository.KakaoLoginRepositoryImpl
 import com.happy.friendogly.data.repository.MemberRepositoryImpl
 import com.happy.friendogly.data.repository.PetRepositoryImpl
@@ -16,7 +15,6 @@ import com.happy.friendogly.data.repository.WoofRepositoryImpl
 import com.happy.friendogly.data.source.AddressDataSource
 import com.happy.friendogly.data.source.AuthDataSource
 import com.happy.friendogly.data.source.ClubDataSource
-import com.happy.friendogly.data.source.FootprintDataSource
 import com.happy.friendogly.data.source.KakaoLoginDataSource
 import com.happy.friendogly.data.source.MemberDataSource
 import com.happy.friendogly.data.source.PetDataSource
@@ -25,7 +23,6 @@ import com.happy.friendogly.data.source.WoofDataSource
 import com.happy.friendogly.domain.repository.AddressRepository
 import com.happy.friendogly.domain.repository.AuthRepository
 import com.happy.friendogly.domain.repository.ClubRepository
-import com.happy.friendogly.domain.repository.FootprintRepository
 import com.happy.friendogly.domain.repository.KakaoLoginRepository
 import com.happy.friendogly.domain.repository.MemberRepository
 import com.happy.friendogly.domain.repository.PetRepository
@@ -64,7 +61,6 @@ import com.happy.friendogly.remote.api.BaseUrl
 import com.happy.friendogly.remote.di.RemoteModule
 import com.happy.friendogly.remote.source.AuthDataSourceImpl
 import com.happy.friendogly.remote.source.ClubDataSourceImpl
-import com.happy.friendogly.remote.source.FootprintDataSourceImpl
 import com.happy.friendogly.remote.source.MemberDataSourceImpl
 import com.happy.friendogly.remote.source.PetDataSourceImpl
 import com.happy.friendogly.remote.source.WoofDataSourceImpl
@@ -92,6 +88,7 @@ class AppModule(context: Context) {
     private val clubService =
         RemoteModule.createClubService(
             baseUrl = baseUrl,
+            localModule = localModule,
             tokenManager = tokenManager,
             authenticationListener = authenticationListener,
         )
@@ -129,8 +126,6 @@ class AppModule(context: Context) {
     private val addressDataSource: AddressDataSource =
         AddressDataSourceImpl(addressModule = addressModule)
     private val kakaoLoginDataSource: KakaoLoginDataSource = KakaoLoginDataSourceImpl()
-    private val footprintDataSource: FootprintDataSource =
-        FootprintDataSourceImpl(service = footprintService)
     private val woofDataSource: WoofDataSource = WoofDataSourceImpl(service = woofService)
     private val memberDataSource: MemberDataSource = MemberDataSourceImpl(service = memberService)
     private val petDataSource: PetDataSource = PetDataSourceImpl(service = petService)
@@ -141,8 +136,6 @@ class AppModule(context: Context) {
     private val tokenRepository: TokenRepository = TokenRepositoryImpl(source = tokenDataSource)
     private val kakaoLoginRepository: KakaoLoginRepository =
         KakaoLoginRepositoryImpl(dataSource = kakaoLoginDataSource)
-    private val footprintRepository: FootprintRepository =
-        FootprintRepositoryImpl(source = footprintDataSource)
     private val woofRepository: WoofRepository = WoofRepositoryImpl(source = woofDataSource)
     private val memberRepository: MemberRepository = MemberRepositoryImpl(source = memberDataSource)
     private val petRepository: PetRepository = PetRepositoryImpl(source = petDataSource)
@@ -166,7 +159,7 @@ class AppModule(context: Context) {
     val deleteTokenUseCase: DeleteTokenUseCase =
         DeleteTokenUseCase(repository = tokenRepository)
     val getFootprintInfoUseCase: GetFootprintInfoUseCase =
-        GetFootprintInfoUseCase(repository = footprintRepository)
+        GetFootprintInfoUseCase(repository = woofRepository)
     val postFootprintUseCase: PostFootprintUseCase =
         PostFootprintUseCase(repository = woofRepository)
     val getNearFootprintsUseCase: GetNearFootprintsUseCase =
