+++ android/app/src/main/java/com/happy/friendogly/data/source/WoofDataSource.kt
@@ -1,6 +1,7 @@
 package com.happy.friendogly.data.source
 
 import com.happy.friendogly.data.model.FootprintDto
+import com.happy.friendogly.data.model.FootprintInfoDto
 import com.happy.friendogly.data.model.FootprintMarkBtnInfoDto
 import com.happy.friendogly.data.model.FootprintSaveDto
 import com.happy.friendogly.remote.model.request.FootprintRequest
@@ -14,4 +15,6 @@ interface WoofDataSource {
         latitude: Double,
         longitude: Double,
     ): Result<List<FootprintDto>>
+
+    suspend fun getFootprintInfo(footprintId: Long): Result<FootprintInfoDto>
 }
