--- android/app/src/main/java/com/happy/friendogly/domain/repository/FootprintRepository.kt
@@ -1,7 +1,0 @@
-package com.happy.friendogly.domain.repository
-
-import com.happy.friendogly.presentation.ui.woof.model.FootprintInfo
-
-interface FootprintRepository {
-    suspend fun getFootprintInfo(footprintId: Long): Result<FootprintInfo>
-}
