+++ android/app/src/main/java/com/happy/friendogly/domain/repository/WoofRepository.kt
@@ -1,6 +1,7 @@
 package com.happy.friendogly.domain.repository
 
 import com.happy.friendogly.presentation.ui.woof.model.Footprint
+import com.happy.friendogly.presentation.ui.woof.model.FootprintInfo
 import com.happy.friendogly.presentation.ui.woof.model.FootprintMarkBtnInfo
 import com.happy.friendogly.presentation.ui.woof.model.FootprintSave
 
@@ -16,4 +17,6 @@ interface WoofRepository {
         latitude: Double,
         longitude: Double,
     ): Result<List<Footprint>>
+
+    suspend fun getFootprintInfo(footprintId: Long): Result<FootprintInfo>
 }
