+++ android/app/src/main/java/com/happy/friendogly/domain/usecase/GetFootprintInfoUseCase.kt
@@ -1,8 +1,8 @@
 package com.happy.friendogly.domain.usecase
 
-import com.happy.friendogly.domain.repository.FootprintRepository
+import com.happy.friendogly.domain.repository.WoofRepository
 import com.happy.friendogly.presentation.ui.woof.model.FootprintInfo
 
-class GetFootprintInfoUseCase(private val repository: FootprintRepository) {
+class GetFootprintInfoUseCase(private val repository: WoofRepository) {
     suspend operator fun invoke(footprintId: Long): Result<FootprintInfo> = repository.getFootprintInfo(footprintId)
 }
