+++ android/app/src/main/java/com/happy/friendogly/remote/api/ApiClient.kt
@@ -10,6 +10,7 @@ class ApiClient {
     object Footprints {
         private const val BASE_URL = "/footprints"
         const val POST_FOOTPRINT = BASE_URL
+        const val POST_FOOTPRINTS_WALK_STATUS = "$BASE_URL/walk-status"
         const val GET_FOOTPRINTS_NEAR = "$BASE_URL/near"
         const val GET_FOOTPRINT_INFO = "$BASE_URL/{footprintId}"
         const val GET_FOOTPRINT_MINE_LATEST = "$BASE_URL/mine/latest"
