+++ android/app/src/main/java/com/happy/friendogly/remote/api/WoofService.kt
@@ -2,12 +2,14 @@ package com.happy.friendogly.remote.api
 
 import com.happy.friendogly.remote.model.request.FootprintRequest
 import com.happy.friendogly.remote.model.response.BaseResponse
+import com.happy.friendogly.remote.model.response.FootprintInfoResponse
 import com.happy.friendogly.remote.model.response.FootprintMarkBtnInfoResponse
 import com.happy.friendogly.remote.model.response.FootprintSaveResponse
 import com.happy.friendogly.remote.model.response.FootprintsNearResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
+import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface WoofService {
@@ -24,4 +26,9 @@ interface WoofService {
 
     @GET(ApiClient.Footprints.GET_FOOTPRINT_MINE_LATEST)
     suspend fun getFootprintMarkBtnInfo(): BaseResponse<FootprintMarkBtnInfoResponse>
+
+    @GET(ApiClient.Footprints.GET_FOOTPRINT_INFO)
+    suspend fun getFootprintInfo(
+        @Path("footprintId") footprintId: Long,
+    ): BaseResponse<FootprintInfoResponse>
 }
